Added some of the newer countries and fixes
Netherlands Antilles (AN) has been split in to new countries but has
been left in for compatibility with some shippers. Probably needs
removing in the future.

Canary Islands (IC) has been added despite not being a official code
only reserved. The 3 letter code ICA is not official or reserved. Canary
Islands (IC) is used by shippers and although part of Spain is not part
EU VAT region so can not be under ES.

Isle of Mann (IM), although a country, has been added under the United
Kingdom as it shares the EU VAT prefix GB and VAT numbers with the
United kingdom. It also shares postcodes and shippers treat as part of
GB. Probably needs adding as a county in the future and translating IM
to GB in the shipping modules and VAT validation.