Update stylesheet.css
Removed white space and added a Reset class at the top which works with OpenCart 1.5.5.1 and has worked for me since 1.4 - the reset is a default standard CSS reset. I have also made the coded more one lined and removed extra ";" - the code is now from 32kb and compressed to 

Compressed CSS from 32kb to 27kb

Code is more minimal and cleaner.

- removed #dddddd to #ddd, likewise with other hex codes
- added a CSS reset I have been using since v1.4 and is cross browser supported
- removed extra lines and indentation to support one liners - small file size
- removed extra ";" to make code more readable and faster
- removed padding: 5px 5px 5px 5px style commands and reduced to padding:5px
- renders 100% on Google Page speed
- css changes:

 -- #container is now margin:0 auto rather than margin-left:auto;margin-right:auto;
 -- added strong as b for bold is obsolete since xHTML 4

Nicer code with no changes to the main design.