Added comments to the header.tpl and footer.tpl
Comments are much needed in the default theme, even the dumber ones,
because it is used as a framework for people to build they own theme or
just for modifications. Everything possible should be commented. I'm
using this logic (for html) as I find more clear to visualize what was
done also it has a lot of adopters. All the important "blocks" should be
commented this way:
<!-- header -->
<header></header>
<!-- /header -->
Easy to visualize and understand. If it's approved, I can keep
"documenting" the theme.