Update PP Express to Handle Lengthy Product Descriptions
Upon checkout, some of our customers were receiving error 11812 (the value of description parameter has been truncated.) This caused quite a mess since the orders weren't showing up in the admin and the customer wasn't receiving a confirmation. To address this, I made the described change so that the description adheres to the 127 single-byte character limit that has been imposed by Paypal.