Fixed an error that was causing profiles to be cancelled when you viewed them. Caused by the controller being loaded to a variable and executing the cancel call.
Changed the return of a method to give an array of actions available by the module.

Not happy that it is using a controller method (should be a model method) but cannot "try" to load a model without an error currently. Model loaded maybe needs to be changed to return false. Will look at this. todo.