Changes quantity Input from Normal Text to Number
Changes the Quantity input box to a number instead of a text.  This added a Up and Down Arrow to the box on a normal browser (except older ones that don't support number as an input, they default to text), also brings up a number pad on a mobile device instead of the normal text input keyboard.

"#product input[type=\'number\']," is added to line 437 to make the ajax query work with the new number input.
DO NOT REMOVE THE TEXT INPUT, This will cause older browsers like IE6 to only add 1 item, even if the input is 20 as they default to text.