Pass all translations to view in Controller::render()
There are tons of code as following in Controller derived classes,

	$this->data['head_title'] = $this->language->get('head_title');
	$this->data['......'] = $this->language->get('......');

it's boring and not insteresting. we can avoid this is by adding one line of code
in Controller::render(),

	extract($this->language->all());

where all() is a new method of Language class, it just returns all translations.

By doing this, the tons of boring code can be removed without any side effect.