Text changes - Consistency Patch - Gift Vouchers
Removal of the "Gift Certificate" and "Voucher" terminology everywhere
in favor of "Gift Voucher".
This wasn't really based on my preference for voucher. I just wanted us
to say one thing or the other and stick to it, i.e. be consistent, not
mix the two...certificate just drew the short straw. I'm equally fine
with Gift Certificate. But if we call it Gift Certificate then let's
call it Gift Certificate throughout.

There might be a couple of other small changes as Github doesn't let me
pick out specific lines of files to include in patch