Corrected a major bug
Corrected a major bug, which have been around for years.

Since the type attribute on the password input where set to "text", the password where visible, also allowing computers, tablets, smartphones and so on to enter the password to the dictionary, which is a standard action for most mobile devices. Computers would remember the password and possible suggest that in the future.

Again, this "bug" have been around for way to long now, and it kinda piss me of, especially since it is a major security risk, and that you can compromise your (possible go-to) password.

Please do pull in this pull request, so that we can get rid of this bug.