Enable Facebook Social Plugins
There is a wide range of popular modules that insert Facebook plugins
into OpenCart shops.

Trouble is, all of them need to include the SDK in the template,
especially if they go down the HTML5 or FBML route.

One issue is that just to add a module, there needs to be a
modification.

Second issue is that if multiple such extensions are used, each adding
the Facebook Javascript, it can lead to conflicts. Serious conflicts.

This patch adds the Facebook SDK Javascript into the Default theme
supplied with core. Thus, the extensions now only need to worry about
the module, and dont need to do any modifications or worry about whether
FB JS has already been loaded.

Would be useful for a large proportion of OpenCart shops.