Fixed Reset Button Glitch
When you get a shipping module information, the main ajax complete() function execute before the success() function in the child ajax call, thus, when you press the "next" button on checkout page, you will se a "Loading" message as expected, but the button will go to the reset state ("Next") before all the shipping method data appear. Most of the time this is not a problem if you dealing with fixed data for shipping, but if your shipping module call a website in order to get real onlines prices, this may took almost 5 seconds, so the user will experiment a 5 seconds delay between the reset state and the update of the shipping data.
The proposed change add an extra wait, that rely on the child ajax call in order to opdate the button to the reset state, so the user now will see the "Loading" state of the button all the time until shipping data is displayed. If any error occur, the system will work as usual.