Installer index missing broad HTTPS switch
OC does not install in a variety of modern stacks that include anything besides standard headers using standard flags. This looks for other headers using any of the 5 common flags, then rewrites HTTPS to true and adds new index for PROTOCOL. This is reflected in startup.php with the addition of a similar proxy IP rewrite.