New HTTPS and forwarded IP rewrites
There needs to be a more encompassing mechanism for rewriting both `$_SERVER['HTTPS']` and `$_SERVER['REMOTE_ADDR']` indices for later use. Also there should be a $_SERVER['PROTOCOL'] for a variety of switches or use cases when relative url (or relative URi) wont cut it. The issue with the old rewrite was that it only looked for `HTTP_X_FORWARDED_PROTO` and `HTTP_X_FORWARDED_SSL` -- depending on server instance, there could be more headers with other non-standard values. This looks for them all containing any value, breaks, switched HTTPS true, on we go.