Use Lower Price of Specials and Discounts
If you have a Special Price for a product it completely ignores the Discount Price

Example:
Base Price is $10/ea
Special Price is $8/ea
Buy 4 or More is $6/ea

The special price of $8 overrides the $6 lower quantity discount because of the order of execution.  This addition of code will choose the lower price of the two and give that price.