Search Anywhere In Name or Model
This makes searching within a product name or model easier and doesn't require you to add on the % wildcard to the beginning of your search term when you don't know the exact way a product is named.  When you have 50,000+ products in a catalog you tend to only remember parts of their names.

It still works as it did if you begin typing the full product name.