Site Search by Partial SKU, Model, UPC, etc.
Allows for customers searching on the site to only know partial SKU's, Models, UPC's, etc.

This is useful when you may have your Model set to something similar to the company SKU, but not exactly.  It also doesn't assume that a customer knows exactly what they are looking for all of the time.