Update order.php
An artifact was showing up in the order account > order history > order information right after Order ID:
```
#%s
```
Now, I'm unfamiliar with twig and it's quirks but to me, it seemed as if maybe it was attempting to pull in a missing variable so I looked in the controller and noticed text_order_id wasn't defined and after defining it the artifact disappeared.

Also, I noticed that
```
 $data['text_invoice_no'] = $this->language->get('text_invoice_no');
```
has been removed from github.
Why?  Am I missing something?