Fix email delivery
Creates a centralized email delivery function. This function does most of the leg work in sending the email through the mail library.

It will attempt to auto correct email headers for from, reply_to, cc, to.

Issue arises when sending SMTP email. Most providers will not allow the FROM header to be anything other than the logged in SMTP user. Thus this patch sets the FROM to be that of the SMTP username if it qualifies as a full email (as most services require). The reply_to header is automatically set to the original FROM header.