Fixed PayPal Express Refund feature in the order admin area. Seems like a lot of different minor variable changes have crept into the code causing these issues - seems to be last fragments of raw PHP in Twig files.
Increased the version number from the expected "3.0.2.1_rc" to "3.0.3.0_rc" - this is due to the rework of the PP Express code and that it has added new features requiring some settings to be updated by the merchant (i.e. not just a bug fix)

Updated the PayPal Express Module to the new v4 PayPal button system which includes PayPal, Credit, Cards, Giro and Venmo (US).

Merchants will need to update the preference of button layouts and re-save PP Express settings after upgrade.

Squashed commits to keep this code update together as similar update will be required on the master shortly.