Update common.js
When option elements are added, the list box disappears on the focus event. If auto-completion of forms is enabled in the Chrome browser (v102.0.5005.63), then a window with options from the browser will appear for a while and disappear when new option elements are added.

Now the disappearance will not occur.
When you first click on the field, no windows will appear until you start typing or click the field a second time.