Added the rest of mandatory subscription fields
These fields needs to be added in the array in order to complete the transaction with the addTransaction() method as to fill the mandatory fields on the database. The subscription_id is not included because we are already mentoring the user to lookup the subscription prior to add it to the subscription transaction table.