Added subscription status lookup in addTransaction
This validation would also be a good addition in the cron/subscription as a subscription status total from the catalog. However, in the addTransaction() method, this would indeed be needed since, not only it doesn't validate anywhere else in the sale/subscription controller but also that the config set can be manipulated from the browser as opposed to the database.