Removed trial_remaining field
Does not exist on the database. As per this commit solution: https://github.com/opencart/opencart/pull/11994/commits/9b2e9af7ce0b65309491ecffdfe78969fb7541fe , the $remaining value must only be passed within the addSubscriptionTransaction() method in order for an extension to take care of the rest with the specified value of 0 when promotional features are involved. Therefore, since the cron/subscription controller is unable to identify the trial_remaining and remaining field, the cron cycle will never complete without this fix.

Furthermore, the trial_duration and duration fields do need to be above 0 since the trial_duration in the admin subscription plans validate method is already verified. Therefore, validating a trial_duration and a duration of a value 0 to return a trial amount or a full subscription's amount would not return the expected results, since the amount must be validated to a maximum extent of a year, not as an unlimited value as, again, the admin subscription plans already validates the trial_duraiton field on anyhow.