Added missing removal fields in upgrade/upgrade_8
The dates have now been restored in opencart-3. In OC 4.0.2.2 release, the subscription database fields have changed since previous OC 4 releases but the previous fields for the upgrade haven't been queried for removal.

If they're not added, the database fields won't be balanced when inserting new data into the database.