Update download.php - Text update
Update to show that big letters are also allowed.

In the next step there is the question with the uploads itself. It is possible to upload filenames with spaces ... 
If there is the possibility then there should be a error message that the filename format is not correct before the file is unloaded.