Update language.php
The problem: when a cookie is already set (e.g. de-de) and the installation is made fresh with English only, each database call for entries (e.g. stock status, order status, returns, etc.) is NOT successful!
Because then the language id in config_language_id is always 0.
Therefore the db result is always an empty array!

This commit solves this problem, or define the path cookies as it is done already for the OCSESSID