Button titles in the Extension Installer
Now we have the button called Install in the list of Installed Extensions, then in the modules list again called Install, it's a bit confusing.
In fact, at this stage we are unpacking the archive, so why not call the button Unzip, and in the list of modules leave it Install.