Update installer.php
I changed an xml file and changed its extension to ocmod.zip and went to install. Thereafter, every time I attempted to load the extension installer page, I was greeted with a PHP unauthorised zip file error and the page just wouldn't load. This update checks for errors with the zip files, if there's an error it will delete the file and continue through the loop without crashing the page