Add OPTIMIZE TABLE to db.php
With standard MySQL operation, InnoDB does not free up disk space when deleting records. In fact, it is marked "Deleted", but there is no actual deletion.
To free up space, you need to perform:
OPTIMIZE TABLE foo;

This will recreate the table and free up space.

https://dev.mysql.com/doc/refman/8.4/en/optimize-table.html#optimize-table-innodb-details

Adding optimization will help save space on the server over a long period of time and speed up work so that the table does not grow infinitely in size.