feat(docker): add intl extension and sort php extensions
This commit adds the `intl` PHP extension to the Docker image. Although it is not a strict dependency, the extension is beneficial as the `oc_validate_email()` function may leverage the `idn_to_ascii()` function for more robust email validation.

The list of installed extensions is also sorted alphabetically to improve readability and consistency.

Signed-off-by: Anton Semenov <20430159+trydalcoholic@users.noreply.github.com>